requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
gunicorn
psycopg2-binary
python-dotenv
werkzeug
textract
docconverter
beautifulsoup4
lxml
playwright
markdown2
docx2txt
aiohttp